test: timeRange helper functions
