+++ src/helpers/timeRange.js
@@ -1,13 +1,3 @@
-const d1 = new Date(2019, 6, 1);
-const d2 = new Date(2018, 6, 1);
-const d3 = new Date(2019, 8, 1);
-
-const objects = [
-  { orderDate: d1, name: 'Test 1' },
-  { orderDate: d2, name: 'Test 2' },
-  { orderDate: d3, name: 'Test 3' },
-];
-
 const sortDates = (array) => {
   return array.sort((a, b) => {
     return a.orderDate - b.orderDate;
@@ -33,9 +23,3 @@ const createArrayMonths = (firstMonth, periods) => {
 };
 
 export { sortDates, monthsDiff, createArrayMonths };
-
-const sortedDates = sortDates(objects);
-const firstMonth = sortedDates[0].orderDate;
-const lastMonth = sortedDates[sortedDates.length - 1].orderDate;
-const periods = monthsDiff(firstMonth, lastMonth);
-const arrayMonths = createArrayMonths(firstMonth, periods);
